Changes to shipping costs and allowing orders that have full discount on the products. Open to read more.
This package was using the total of the basket without having consideration for the shipping costs. So, if a basket was totalled at 0 cost because of discount but still had a shipping cost, it wouldn't allow to place the order.
Now you can do it.
Also, it was passing shipping discount information as a product, which was very confusing. Now it's using the aptly named SHIPDISCAMT for that.
Dangerous commit that only works well with custom code we made for our own shops.