refs #36361: Allowing orders where the discount value is higher than the basket value
This is necessary to ensure that you don't have a situation where you have a higher value of discounts in comparison to basket value.
Example: Product1: 110€, Discount1: 50€, Discount2: 70€. Total discount: 120€
This code ensures that the end result on the paypal bill is like this:
Product1: 110€
Discount1: 50€
Discount2: 60€ (Even though it's actually 70€)
Total: 0€
The reason we need this is because paypal doesn't have a place for basket discounts (only shipping discount).
So we need to put these discount values in lines that are intended for products.
Meaning that in the end, paypal will sum all product and discount values and the result will have to be equal to the total of the order.